test: fix more type mismatches and signature test

+++ tests/integration/api_notification_test.go
@@ -155,7 +155,7 @@ func TestAPINotification(t *testing.T) {
 		AddTokenAuth(token)
 	resp = MakeRequest(t, req, http.StatusOK)
 	DecodeJSON(t, resp, &newStruct)
-	assert.Equal(t, 0, newStruct.New)
+	assert.Equal(t, int64(0), newStruct.New)
 }
 
 func TestAPINotificationPUT(t *testing.T) {
